+++ src/main/java/com/robertx22/age_of_exile/aoe_data/GeneratedData.java
@@ -17,6 +17,7 @@ import com.robertx22.age_of_exile.aoe_data.database.perks.Perks;
 import com.robertx22.age_of_exile.aoe_data.database.perks.SpellModifierPerks;
 import com.robertx22.age_of_exile.aoe_data.database.runes.Runes;
 import com.robertx22.age_of_exile.aoe_data.database.runewords.Runewords;
+import com.robertx22.age_of_exile.aoe_data.database.salvage_outputs.SalvageOutputsAdder;
 import com.robertx22.age_of_exile.aoe_data.database.spells.Spells;
 import com.robertx22.age_of_exile.aoe_data.database.stats.AutoDatapackStats;
 import com.robertx22.age_of_exile.aoe_data.database.stats.DatapackStatAdder;
@@ -69,6 +70,7 @@ public class GeneratedData {
             new AutoDatapackStats().registerAll();
 
             new FavorAdder().registerAll();
+            new SalvageOutputsAdder().registerAll();
         }
     }
 }