+++ src/main/java/com/robertx22/age_of_exile/config/forge/GearSalvageConfig.java
@@ -1,54 +1,13 @@
 package com.robertx22.age_of_exile.config.forge;
 
-import com.robertx22.age_of_exile.saveclasses.item_classes.GearItemData;
-import com.robertx22.age_of_exile.uncommon.utilityclasses.ItemUtils;
-import com.robertx22.age_of_exile.uncommon.utilityclasses.RandomUtils;
-import me.sargunvohra.mcmods.autoconfig1u.annotation.ConfigEntry;
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.Items;
-
-import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.List;
-
 public class GearSalvageConfig {
-
-    public int BASE_ITEM_VALUE = 500;
-
-    public int VALUE_ADDED_PER_AFFIX = 100;
-    public int VALUE_ADDED_PER_SOCKET = 50;
-    public int VALUE_ADDED_PER_FULL_SOCKET = 150;
-
-    @ConfigEntry.Gui.CollapsibleObject
-    public LinearScalingConfig VALUE_PER_LEVEL_SCALING = new LinearScalingConfig(0.02D);
-
-    List<ItemAndPrice> SALVAGE_RESULTS = new ArrayList<ItemAndPrice>() {
-        {
-            // add(new ItemAndPrice(500, ModRegistry.MISC_ITEMS.MAGIC_ESSENCE, 1000));
-            //add(new ItemAndPrice(700, ModRegistry.MISC_ITEMS.RARE_MAGIC_ESSENCE, 150));
-            add(new ItemAndPrice(500, Items.IRON_INGOT, 50));
-            add(new ItemAndPrice(75, Items.IRON_NUGGET, 25));
-            add(new ItemAndPrice(100, Items.GOLD_NUGGET, 25));
-        }
-    };
 
     public GearSalvageConfig() {
     }
 
-    public List<ItemStack> getResult(GearItemData gear) {
 
-        ItemStack stack = new ItemStack(ItemUtils.randomDustFor(gear.level));
-        stack.setCount(Math.max(1, (int) gear.getRarity()
-            .valueMulti()));
 
-        List<ItemStack> list = new ArrayList<>();
-        list.add(stack);
-
-        return list;
-
-        //return getResultForValue(getValue(gear)); // todo how to make configurable agian..
-    }
-
+    /*
     private int getValue(GearItemData gear) {
 
         int value = BASE_ITEM_VALUE;
@@ -65,28 +24,6 @@ public class GearSalvageConfig {
         return value;
     }
 
-    private List<ItemStack> getResultForValue(int value) {
-        List<ItemStack> stacks = new ArrayList<>();
-
-        int smallest = SALVAGE_RESULTS.stream()
-            .min(Comparator.comparingInt(x -> x.price))
-            .get().price;
-
-        while (value >= smallest) {
-
-            List<ItemAndPrice> list = new ArrayList<>();
-            for (ItemAndPrice x : SALVAGE_RESULTS) {
-                if (x.price <= value) {
-                    list.add(x);
-                }
-            }
-            ItemAndPrice random = RandomUtils.weightedRandom(list);
-            stacks.add(new ItemStack(random.getItem()));
-            value -= random.price;
-        }
-
-        return stacks;
-
-    }
+     */
 
 }